+++ examples/shardingsphere-jdbc-example/sharding-example/sharding-raw-jdbc-example/src/main/java/org/apache/shardingsphere/example/sharding/raw/jdbc/config/ShardingDatabasesAndTablesConfigurationRange.java
@@ -75,7 +75,7 @@ public final class ShardingDatabasesAndTablesConfigurationRange implements Examp
     
     private static Properties getProperties() {
         Properties result = new Properties();
-        result.setProperty("worker.id", "123");
+        result.setProperty("worker-id", "123");
         return result;
     }
 }