+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-common/src/test/java/org/apache/shardingsphere/sharding/algorithm/sharding/mod/HashModShardingAlgorithmTest.java
@@ -42,7 +42,7 @@ public final class HashModShardingAlgorithmTest {
     @Before
     public void setup() {
         HashModShardingAlgorithm shardingAlgorithm = new HashModShardingAlgorithm();
-        shardingAlgorithm.getProps().setProperty("sharding.count", "4");
+        shardingAlgorithm.getProps().setProperty("sharding-count", "4");
         shardingAlgorithm.init();
         shardingStrategy = new StandardShardingStrategy("order_type", shardingAlgorithm);
     }